+++ src/app/[locale]/(main)/company/[id]/page.tsx
