+++ src/app/[locale]/(main)/edit/page.tsx
@@ -4,14 +4,16 @@ import { getTranslations } from "next-intl/server";
 import CompanyForm from "@/app/[locale]/(main)/_components/company-form";
 import MaxWidthWrapper from "@/components/max-width-wrapper";
 import getAllowedCompany from "@/app/actions/get-allowed-company";
+import getCompanyId from "@/lib/get-company-id";
 
 type Props = {
-  params: { locale: string; id: string };
+  params: { locale: string };
 };
 
 export const generateMetadata = async ({
-  params: { locale, id },
+  params: { locale },
 }: Props): Promise<Metadata> => {
+  const id = getCompanyId();
   const t = await getTranslations({ locale, namespace: "Metadata" });
 
   const company = await getAllowedCompany(id);
@@ -27,7 +29,8 @@ export const generateMetadata = async ({
   };
 };
 
-const EditCompany = async ({ params: { id } }: Props) => {
+const EditCompany = async () => {
+  const id = getCompanyId();
   const company = await getAllowedCompany(id);
   if (!company) return notFound();
