+++ src/app/[locale]/(main)/members/_components/delete-member.tsx
