+++ src/app/[locale]/(main)/price-types/_components/delete-price-type.tsx
