+++ src/app/[locale]/(main)/price-types/_components/price-type-form.tsx
