+++ src/app/[locale]/(main)/roles/_components/delete-role.tsx
