+++ src/app/[locale]/(main)/roles/page.tsx
@@ -1,20 +1,19 @@
 import { Metadata } from "next";
+import Link from "next/link";
 import { headers, cookies } from "next/headers";
 import { notFound } from "next/navigation";
 import { getTranslations } from "next-intl/server";
-import type { Prisma } from "@prisma/client";
-import { MoreHorizontal } from "lucide-react";
-import MemberForm from "./_components/member-form";
-import InvitationForm from "./_components/invitation-form";
-import DeleteMember from "./_components/delete-member";
+import type { CompanyRole as RoleType } from "@prisma/client";
+import { MoreHorizontal, Plus, Pencil } from "lucide-react";
+import DeleteRole from "./_components/delete-role";
 import MaxWidthWrapper from "@/components/max-width-wrapper";
 import Pagination from "@/components/ui/pagination";
-import { Button } from "@/components/ui/button";
+import { Button, buttonVariants } from "@/components/ui/button";
 import {
   Card,
+  CardDescription,
   CardHeader,
   CardTitle,
-  CardDescription,
 } from "@/components/ui/card";
 import {
   DropdownMenu,
@@ -22,21 +21,19 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import request from "@/lib/request";
+import { cn } from "@/lib/utils";
 import { PaginationType } from "@/types";
-
-type MemberType = Prisma.CompanyMembersGetPayload<{
-  include: { user: true; companyRole: true };
-}>;
+import getCompanyId from "@/lib/get-company-id";
 
 type Props = {
-  params: { locale: string; id: string };
+  params: { locale: string };
   searchParams: {
     page?: number;
   };
 };
 
 type PaginatedData = {
-  result: MemberType[];
+  result: RoleType[];
   pagination: PaginationType;
 };
 
@@ -46,19 +43,19 @@ export const generateMetadata = async ({
   const t = await getTranslations({ locale, namespace: "Metadata" });
 
   return {
-    title: t("membersTitle"),
-    description: t("membersDescription"),
+    title: t("rolesTitle"),
+    description: t("rolesDescription"),
   };
 };
 
-const getCompanyMembers = async (companyId: string, page: number) => {
+const getCompanyRoles = async (companyId: string, page: number) => {
   try {
     const cookieStore = cookies();
     const headersList = headers();
     const cookie = headersList.get("cookie");
 
     return await request<PaginatedData>(
-      `/api/company/${companyId}/member?page=${page}&limit=10`,
+      `/api/company/${companyId}/role?page=${page}&limit=10`,
       {
         headers: {
           "Accept-Language": cookieStore.get("NEXT_LOCALE")?.value,
@@ -72,13 +69,14 @@ const getCompanyMembers = async (companyId: string, page: number) => {
   }
 };
 
-const Members = async ({ params: { id }, searchParams: { page } }: Props) => {
-  const data = await getCompanyMembers(id, page ?? 1);
+const Roles = async ({ searchParams: { page } }: Props) => {
+  const id = getCompanyId();
+  const data = await getCompanyRoles(id, page ?? 1);
   if (!data) return notFound();
 
   const { result, pagination } = data;
 
-  const t = await getTranslations("Member");
+  const t = await getTranslations("Role");
 
   return (
     <MaxWidthWrapper className="mb-8">
@@ -87,18 +85,22 @@ const Members = async ({ params: { id }, searchParams: { page } }: Props) => {
           <h1 className="text-4xl font-bold tracking-tight">{t("title")}</h1>
           <p className="text-lg text-muted-foreground">{t("subtitle")}</p>
         </div>
-        <InvitationForm companyId={id} />
+        <Link
+          href={`/new-role`}
+          className={cn(buttonVariants({ variant: "default" }), "space-x-2")}
+        >
+          <Plus />
+          <span>{t("new")}</span>
+        </Link>
       </div>
       <div className="mt-4">
         {result.length > 0 && (
           <div className="grid grid-flow-row auto-rows-max place-items-center grid-cols-1 gap-4 pt-4 md:place-items-start md:grid-cols-2">
-            {result.map((member) => (
-              <Card key={member.userId} className="h-full w-full">
+            {result.map((role) => (
+              <Card key={role.id} className="h-full w-full">
                 <CardHeader>
                   <div className="flex flex-row justify-between">
-                    <CardTitle className="text-xl truncate">
-                      {member.user.name || member.user.email}
-                    </CardTitle>
+                    <CardTitle>{role.name}</CardTitle>
                     <DropdownMenu>
                       <DropdownMenuTrigger asChild>
                         <Button variant="ghost" className="h-8 w-8 p-0">
@@ -110,27 +112,28 @@ const Members = async ({ params: { id }, searchParams: { page } }: Props) => {
                         align="end"
                         className="flex flex-col"
                       >
-                        <MemberForm companyId={id} member={member} />
-                        <DeleteMember companyId={id} memberId={member.userId} />
+                        <Link
+                          href={`/edit-role/${role.id}`}
+                          className={buttonVariants({ variant: "ghost" })}
+                        >
+                          <Pencil className="mr-2 h-4 w-4" />
+                          <span>{t("edit")}</span>
+                        </Link>
+                        <DeleteRole companyId={id} roleId={role.id} />
                       </DropdownMenuContent>
                     </DropdownMenu>
                   </div>
-                  <CardDescription>{`${t("role")}: ${
-                    member.companyRole.name
-                  }`}</CardDescription>
+                  <CardDescription>{`${t("id")}: ${role.id}`}</CardDescription>
                 </CardHeader>
               </Card>
             ))}
           </div>
         )}
-        <Pagination
-          href={`/company/${id}/members?page=`}
-          pagination={pagination}
-        />
+        <Pagination href={`/roles?page=`} pagination={pagination} />
         <div />
       </div>
     </MaxWidthWrapper>
   );
 };
 
-export default Members;
+export default Roles;
