+++ src/app/[locale]/(main)/stocks/_components/delete-stock.tsx
