+++ src/app/[locale]/(main)/stocks/_components/stock-form.tsx
