+++ src/lib/get-company-id.ts
@@ -1,0 +1,3 @@
+export default function getCompanyId() {
+  return "bestsoft";
+}
